Remove `expect/actual` arrangement for the `LoggingScope` class
